refactor activity type selectors
closes #438
