+++ projects/ngx-d3timeline/src/lib/options/selectors/activity-options.selectors.ts
@@ -1,13 +1,70 @@
 import { createSelector } from '../../store-lib/selector/create-selector';
 import { selectOptions } from '../../store/state';
 import { MemoizedSelector } from '../../store-lib/selector/memoized-selector';
+import { selectTypeOptions } from './type-options.selectors';
+import { partial1, partial2 } from '../../core/partial';
+import { getTypeActivityOptions } from '../options-utils';
+import { ActivityOptions } from '../options';
 
-export const selectActivityOptions = createSelector(
+const selectActivityOptions = createSelector(
   selectOptions,
   options => options.activity
 );
 
-export const selectGetGlobalActivityOption = <T>(
+const selectTypeActivityOptions = createSelector(
+  selectTypeOptions,
+  partial1(getTypeActivityOptions)
+);
+
+const selectGetTypeActivityOption = <T>(
   key: string
-): MemoizedSelector<T> =>
+): MemoizedSelector<(type: string) => T> =>
+  createSelector(
+    selectTypeActivityOptions,
+    options =>
+      getTypeActivityOption.bind(null, options, key) as (type: string) => T
+  );
+
+function getTypeActivityOption<T>(
+  typeActivityOptions: (type: string) => ActivityOptions,
+  key: string,
+  type: string
+): T {
+  return typeActivityOptions(type) && typeActivityOptions(type)[key];
+}
+
+const selectGetGlobalActivityOption = <T>(key: string): MemoizedSelector<T> =>
   createSelector(selectActivityOptions, options => options[key]);
+
+const selectGetActivityOption = <T>(
+  key: string
+): MemoizedSelector<(type: string) => T> =>
+  createSelector(
+    selectGetTypeActivityOption<T>(key),
+    selectGetGlobalActivityOption<T>(key),
+    partial2(getActivityOption)
+  );
+
+function getActivityOption<T>(
+  typeActivityOption: (type: string) => T,
+  globalActivityOption: T,
+  type: string
+): T {
+  return typeActivityOption(type) || globalActivityOption;
+}
+
+export const selectGetActivityLateralMargin = selectGetActivityOption<number>(
+  'lateralMargin'
+);
+export const selectGetActivityDisableDrag = selectGetActivityOption<boolean>(
+  'disableDrag'
+);
+export const selectGetActivityFontFace = selectGetActivityOption<string>(
+  'fontFace'
+);
+export const selectGetActivityFontSize = selectGetActivityOption<number>(
+  'fontSize'
+);
+export const selectGetActivityPadding = selectGetActivityOption<number>(
+  'padding'
+);
